+++ python/packages/brilliant_ble/src/brilliant_ble/__init__.py
@@ -6,5 +6,6 @@ __all__ = ["brilliant_ble"]
 from .brilliant_ble import BrilliantBle
 from .brilliant_ble import BrilliantDeviceType
 from .brilliant_ble import OtaError
+from .brilliant_ble import chunk_lua_string
 
-__version__ = "3.1.0"
+__version__ = "3.1.1"